pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "role-unit"
version = "0.1.0"
description = "Role-Unit: Benchmarking LLMs in multi-agent roles on MMLU"
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
  "aiohttp>=3.9",
  "class-registry>=2.1",
  "groq>=0.9",
  "langchain-community>=0.2",
  "loguru>=0.7",
  "numpy>=1.26",
  "openai>=1.0",
  "pandas>=2.0",
  "python-dotenv>=1.0",
  "pyyaml>=6.0",
  "requests>=2.31",
  "setuptools>=68",
  "shortuuid>=1.0",
  "tenacity>=8.0",
  "tiktoken>=0.7",
  "torch>=2.6.0",
  "wikipedia>=1.4",
]

[dependency-groups]
dev = [
  "pytest>=8.0",
  "ruff>=0.5",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["MAR*"]
